requirements.txt
Flask
opencv-python
numpy
Pillow
svgwrite
gunicorn
requests